dtype consistency

+++ src/diffoptics/optics/Window.py
@@ -155,7 +155,7 @@ class Window(BaseOptics):
         mu = self.n_glass / self.n_ext
         tmp = 1 - mu ** 2 * (1 - (dot_product(window_normal, directions)) ** 2)
         c = dot_product(window_normal, directions)
-        direction_refracted_rays = torch.empty((tmp.shape[0], 3), device=origins.device)
+        direction_refracted_rays = torch.empty((tmp.shape[0], 3), device=origins.device, dtype=origins.dtype)
         direction_refracted_rays[:, 0] = torch.sqrt(tmp) * window_normal[:, 0] + mu * (
                     directions[:, 0] - c * window_normal[:, 0])
         direction_refracted_rays[:, 1] = torch.sqrt(tmp) * window_normal[:, 1] + mu * (
